Extend DatabaseObject class
Implement a subclass resource, that extends the DatabaseObject to be a qualified OAuth resource so a database object can belong to a user and be either public or private.
