+++ openAPE/API/src/main/java/org/openape/api/taskcontext/TaskContext.java
@@ -22,16 +22,16 @@ import java.util.List;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
-import org.openape.api.DatabaseObject;
 import org.openape.api.Property;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.openape.api.Resource;
 
 /**
  * Task context object defined in 7.3.1
  */
 @XmlRootElement
-public class TaskContext extends DatabaseObject {
+public class TaskContext extends Resource {
     private static final long serialVersionUID = 3325722856059287182L;
 
     /**
